Create Character object declaration.js
This code declares and defines a global object, D20_Character, from which all actual characters in the game inherit default values and many important methods. This code is loosely coupled to the Roll20 game engine: the D20_Character object contains load_fromRoll20Character() and load_fromRoll20Token() methods retrieve data from the game. (Corresponding save_ routines have not been coded yet.) All other code authored by me for use with Pathfinder in the Roll20 engine will use this object format; this allows my code to be easily given away to other Storytellers in the community, because all they will have to change is the names of the Roll20 Attributes listed the load and save methods (which refer to Attribute names on the custom character sheet.)

 Also, this code block defines a hash table named characterData, which is a closure with private data and public get() and set() methods. The named properties of this private object correspond to the unique IDs of the Roll20 Characters and Tokens, minus the leading hyphen. (Both Roll20 Characters and Roll20 Tokens are translated into D20_Character objects. Tokens inherit their values from prototype Character objects, and have additional properties, such as xPos and yPos.) This has table is maintained using Lazy Initialization: Roll20 Character and Token objects are not loaded into D20_Characters until they are needed and/or changed.