Create Spell object declarations.js
This code declares and defines a global object, D20_Spell, from which all spell effects in the game inherit default values and many important methods. Caster and target(s) of the spells are defined in terms of my D20_Character object, and the output of the spell's effects are constructed using the create_HTMLDisplay() method. Next, a global variable is defined named spellData; then, all the spells defined in the game so far are loaded as descendants of D20_Spell, and named properties of the spellData object. All spell data is exactly as it appears in the Pathfinder online reference compendium unless otherwise stated.

The following spells have been coded so far:
    Bane
    Bless
    Charm Person
    Color Spray
    Entangle
    Hold Person
    Hunter's Mercy (from the Forgotten Realms campaign setting)
    Sleep