Create Main API (Roll20 event handlers).js
This source file contains the event handler for the on('chat:message') event, which is link between the Roll20 gaming environment's User Interface, and the JavaScript API scripts written by the Game Master. Chat message events which are preceded by an exclamation mark signal an API call, which is then parsed into a command and several parameters, which include the ID of the selected Token, as well as the ID of the targeted Token, if applicable. 

The following API commands have been defined:
    attack (one D20_Character attack another, using one of three weapons being wielded)
    attack_mode (toggle on or off special attack modes, such as fighting defensively and full-round attack actions)
    cast (cast a spell - invokes the onCast() method of the appropriate D20_Spell object)
    clear (clears tint, aura, and status markers from selected Roll20 token)
    export (JSON.stringify() the selected D20_Character)
    finish_effect (completes casting a spell or using an item, once the locator Token is positioned. Invokes the onFinishCast() method.)
    reveal (moves a Token from GMLayer to Object Layer, and outputs their short description to the chat window)
    test_distance (calculate the distance in feet from the selected Token to the targeted one)
    use (use an item - invokes the onUse() method of the appropriate D20_Item object)

This file also contains short event handlers for changes made to Roll20 Tokens, Characters, and Attributes, so that their corresponding D20_Character objects may be updated. Currently the entire D20_Character is re-loaded at once; the code to load (and save) only the Attribute that is needed, is not yet complete.