Create Item object declarations.js
This code declares and defines a global object, D20_Item, from which all actual items in the game inherit default values and many important methods. It also defines a D20_Weapon object which inherits from D20_Item and adds several properties and methods pertaining to functions of a weapon.

This code is loosely coupled to the Roll20 game engine: the D20_Weapon object adds a load_fromRoll20Character() method to retrieve the values for the character's Main hand (melee), Off-hand (melee), and Ranged weapon slots, from the Roll20 custom character sheet.
All other code authored by me for use with Pathfinder in the Roll20 engine will use this object format; this allows my code to be easily given away to other Storytellers in the community, because all they will have to change is the names of the Roll20 Attributes listed the load and save methods (which refer to Attribute names on the custom character sheet.)

Conversely, for specific items (such as Alchemist's Fire) whose properties are to be hard-coded into the game, a database of Items is going to be built. This code block defines a hash table named itemData, which is a closure with private data and public get() and set() methods. The named properties of this private object correspond to the 'name' property of the D20_Item, with spaces and apostrophes removed. These items are loaded into the itemData table as soon as the game data is loaded. Currently, the API does NOT load the values from the 'Item1' section of the Character Sheet. Only Alchemist's Fire has been declared, and this is incomplete.