Create Sarjenka's Roll20 char sheet.html
Custom Character Sheets in Roll20 are the bodies of HTML documents. The names of the inputs correspond to Roll20 Attributes associated with Characters; these are linked and updated automatically. The character sheet is formatted using CSS; outside libraries such as jQuery can NOT be used to change the sheet; however, Roll20 DOES provide some functionality for built-in classes of "repeating fields" (elements of which can be added, modified, and deleted while the game is in progress). Because of this, the sheet does not display well outside of the Roll20 engine.